implementing `ToString` for `Address`
