+++ crates/sdk/std/src/string/traits.rs
@@ -1,12 +1,14 @@
-use crate::{String, StringBuilder, Token};
+use crate::{DecDigit, String, StringBuilder, Token};
 
-/// A trait to be implemented by bytes that want to have a [`String`] representation.
+/// A trait to be implemented by types that want to have a [`String`] representation.
 pub trait ToString {
     /// Returns a [`String`] representing `self`.
     fn to_string(&self) -> String;
 }
 
+/// A trait to be implemented by types that want to have a [`Token`] representation.
 pub trait ToToken {
+    /// Returns a [`Token`] representing `self`.
     fn to_token(&self) -> Token;
 }
 
@@ -15,56 +17,6 @@ impl ToString for bool {
         match *self {
             true => String::new("True"),
             false => String::new("False"),
-        }
-    }
-}
-
-#[repr(transparent)]
-pub struct DecDigit(u8);
-#[repr(transparent)]
-pub struct HexDigit(u8);
-
-impl ToToken for HexDigit {
-    fn to_token(&self) -> Token {
-        let value = self.0;
-        debug_assert!(value < 16);
-
-        if value < 10 {
-            let digit = DecDigit(value);
-            return digit.to_token();
-        }
-
-        // TODO: is that possible to use `seq-macro` here?
-        match value {
-            10 => Token::One(b'A'),
-            11 => Token::One(b'B'),
-            12 => Token::One(b'C'),
-            13 => Token::One(b'D'),
-            14 => Token::One(b'E'),
-            15 => Token::One(b'F'),
-            _ => crate::panic(),
-        }
-    }
-}
-
-impl ToToken for DecDigit {
-    fn to_token(&self) -> Token {
-        let value = self.0;
-        debug_assert!(value < 10);
-
-        // TODO: is that possible to use `seq-macro` here?
-        match value {
-            0 => Token::One(b'0'),
-            1 => Token::One(b'1'),
-            2 => Token::One(b'2'),
-            3 => Token::One(b'3'),
-            4 => Token::One(b'4'),
-            5 => Token::One(b'5'),
-            6 => Token::One(b'6'),
-            7 => Token::One(b'7'),
-            8 => Token::One(b'8'),
-            9 => Token::One(b'9'),
-            _ => crate::panic(),
         }
     }
 }
