+++ crates/sdk/types/Cargo.toml
@@ -11,6 +11,7 @@ description = "Spacemesh Virtual Machine"
 publish = false
 
 [dependencies]
+seq-macro = "0.2.2"
 svm-sdk-std = { path = "../std", default-features = false }
 
 [features]
