+++ crates/sdk/types/src/blob.rs
@@ -1,5 +1,7 @@
 #![allow(unused_must_use)]
 
+use svm_sdk_std::{HexDigit, String, StringBuilder, ToString, ToToken, Token};
+
 macro_rules! impl_blob_type {
     ($ty:ident, $nbytes:expr) => {
         use core::cmp::{Eq, PartialEq};
@@ -133,7 +135,49 @@ macro_rules! impl_blob_type {
         }
 
         impl Eq for $ty {}
+
+        impl ToString for $ty {
+            fn to_string(&self) -> String {
+                let mut sb = StringBuilder::with_capacity("0x".len() + Self::len() * 2);
+                sb.push_token(Token::Two(b'0', b'x'));
+
+                let bytes = self.as_slice();
+                seq_macro::seq!(N in 0..$nbytes {
+                    let byte: u8 = bytes[N];
+
+                    // extracting nibbles
+                    let left = (byte & 0xF0) >> 4;
+                    let right = byte & 0x0F;
+
+                    sb.push_token(HexDigit(left).to_token());
+                    sb.push_token(HexDigit(right).to_token());
+                });
+
+                sb.build()
+            }
+        }
     };
 }
 
 impl_blob_type!(Address, 20);
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn address_to_string() {
+        extern crate std;
+
+        let addr = Address::repeat(0xAB);
+        let string = svm_sdk_std::ToString::to_string(&addr);
+
+        let vec: std::vec::Vec<u8> = string.as_bytes().into();
+        let string = unsafe { std::string::String::from_utf8_unchecked(vec) };
+
+        assert_eq!(
+            string.as_str(),
+            "0xABABABABABABABABABABABABABABABABABABABAB"
+        );
+    }
+}
